Non-functional file parsing implemented hooray

+++ B&B Campaign Assistant/Assets/Engineering/Plugins/GameManager/GameManager/Item.h
@@ -12,7 +12,7 @@ public:
 protected:
 	char *name;			//name
 	float weight;		//weight
-	int price;			//price
+	float price;		//price
 	char *description;	//description of the item
 	float durability;	//number representing the amount of "health" the functional portion of the item has. higher numbers are better. when this reaches 0 the item is unusable
 };
